+++ packages/logic-core/src/data/symbols/houseSymbol.ts
@@ -1,10 +1,10 @@
 import { AnyConfig, ConfigType } from '../config.js';
 import GridData from '../grid.js';
 import { array } from '../dataHelper.js';
-import { Color, State } from '../primitives.js';
-import Symbol from './symbol.js';
+import { Color } from '../primitives.js';
+import NumberSymbol from './numberSymbol.js';
 
-export default class HouseSymbol extends Symbol {
+export default class HouseSymbol extends NumberSymbol {
   public readonly title = 'House';
 
   private static readonly CONFIGS: readonly AnyConfig[] = Object.freeze([
@@ -22,16 +22,24 @@ export default class HouseSymbol extends Symbol {
       description: 'Y',
       configurable: false,
     },
+    {
+      type: ConfigType.Number,
+      default: 2,
+      field: 'number',
+      description: 'Number',
+      explanation: 'Number of houses in this region',
+      configurable: true,
+    },
   ]);
 
   private static readonly EXAMPLE_GRID = Object.freeze(
     GridData.create(['bbbww', 'wwwbw', 'wbbbw', 'wwwww'])
-      .addSymbol(new HouseSymbol(0, 0))
-      .addSymbol(new HouseSymbol(2, 0))
-      .addSymbol(new HouseSymbol(3, 0))
-      .addSymbol(new HouseSymbol(2, 1))
-      .addSymbol(new HouseSymbol(3, 1))
-      .addSymbol(new HouseSymbol(1, 2))
+      .addSymbol(new HouseSymbol(0, 0, 2))
+      .addSymbol(new HouseSymbol(2, 0, 2))
+      .addSymbol(new HouseSymbol(3, 0, 2))
+      .addSymbol(new HouseSymbol(2, 1, 2))
+      .addSymbol(new HouseSymbol(3, 1, 2))
+      .addSymbol(new HouseSymbol(1, 2, 2))
   );
 
   /**
@@ -39,12 +47,14 @@ export default class HouseSymbol extends Symbol {
    *
    * @param x - The x-coordinate of the symbol.
    * @param y - The y-coordinate of the symbol.
+   * @param number - The number of houses in this region.
    */
   public constructor(
     public readonly x: number,
-    public readonly y: number
+    public readonly y: number,
+    public readonly number: number
   ) {
-    super(x, y);
+    super(x, y, number);
   }
 
   public get id(): string {
@@ -52,7 +62,7 @@ export default class HouseSymbol extends Symbol {
   }
 
   public get explanation(): string {
-    return '*Houses* must connect to *exactly one* other house';
+    return '*House numbers* count the number of houses in the region';
   }
 
   public get configs(): readonly AnyConfig[] | null {
@@ -63,20 +73,19 @@ export default class HouseSymbol extends Symbol {
     return HouseSymbol.EXAMPLE_GRID;
   }
 
-  public validateSymbol(grid: GridData): State {
+  public countTiles(grid: GridData): { completed: number; possible: number } {
     const thisX = Math.floor(this.x);
     const thisY = Math.floor(this.y);
-    let complete = true;
     const visited = array(grid.width, grid.height, () => false);
     const connected = array(grid.width, grid.height, () => false);
     const color = grid.getTile(thisX, thisY).color;
-    if (color === Color.Gray) return State.Incomplete;
+    if (color === Color.Gray)
+      return { completed: 0, possible: Number.MAX_SAFE_INTEGER };
     grid.iterateArea(
       { x: thisX, y: thisY },
       tile => tile.color === Color.Gray || tile.color === color,
-      (tile, x, y) => {
+      (_, x, y) => {
         visited[y][x] = true;
-        if (tile.color === Color.Gray) complete = false;
       }
     );
     grid.iterateArea(
@@ -86,27 +95,37 @@ export default class HouseSymbol extends Symbol {
         connected[y][x] = true;
       }
     );
-    let connectedHouses = 0;
+    let completedHouses = 0;
     let possibleHouses = 0;
     for (const symbol of grid.symbols.get(this.id) ?? []) {
-      if (symbol !== this && symbol instanceof HouseSymbol) {
+      if (symbol instanceof HouseSymbol) {
         const symbolX = Math.floor(symbol.x);
         const symbolY = Math.floor(symbol.y);
-        if (connected[symbolY][symbolX]) connectedHouses++;
+        if (connected[symbolY][symbolX]) completedHouses++;
         else if (visited[symbolY][symbolX]) possibleHouses++;
       }
     }
-    if (connectedHouses > 1 || connectedHouses + possibleHouses < 1) {
-      return State.Error;
-    }
-    return complete || (connectedHouses === 1 && possibleHouses === 0)
-      ? State.Satisfied
-      : State.Incomplete;
+    return {
+      completed: completedHouses,
+      possible: completedHouses + possibleHouses,
+    };
   }
 
-  public copyWith({ x, y }: { x?: number; y?: number }): this {
-    return new HouseSymbol(x ?? this.x, y ?? this.y) as this;
+  public copyWith({
+    x,
+    y,
+    number,
+  }: {
+    x?: number;
+    y?: number;
+    number?: number;
+  }): this {
+    return new HouseSymbol(
+      x ?? this.x,
+      y ?? this.y,
+      number ?? this.number
+    ) as this;
   }
 }
 
-export const instance = new HouseSymbol(0, 0);
+export const instance = new HouseSymbol(0, 0, 2);