+++ src/client/instructions/Instruction.tsx
@@ -97,7 +97,7 @@ export default memo(function Instruction({
         ></div>
         <div
           className={cn(
-            'text-center py-1 px-4 flex grow justify-center items-center text-neutral-content',
+            'text-center py-1 px-2 flex grow justify-center items-center text-neutral-content',
             responsive
               ? 'text-sm lg:text-base'
               : large